Remove the 'groups' feature flag
This is deployed and toggled on everywhere, so we can simplify our code
by removing this feature flag.

There's one small wrinkle here, which is that there are still two
distinct designs for the "top bar" -- one for the sidebar and one for
the stream. I've simplified this in this commit, and removed any
reference to groups as a toggle for the two flavours of sidebar, but at
some point we will probably want to unify the two experiences.