Change the search API to search top-level annotations
Fixes #1916.

Get rid of "Message not found" messages on the /stream page.

Problem:

The stream requests the most recent n annotations from the
search API, but what it actually gets is the most recent n annotations
_and replies_. In some cases a reply may be in this set while the parent
annotation that it is a reply to is not.  When that happens, the frontend
just displays "Message not found" in place of this thread (parent
annotation and replies).

Solution:

Change the semantics of the search API: only return the top-level
annotations (not reply annotations) in "rows". Replies are always
filtered out from the Elasticsearch query results. "total" (and "limit" and
"offset") only apply to these top-level annotations.

Additionally, always do a second Elasticsearch query to get _all_ reply
annotations to the matching top-level annotations, and return these in a
new "replies" key in the API response.

This means that whenever the frontend calls the search API, it always
gets back all of the replies to all of the annotations that it gets, so
it always has the complete threads and feeds them all into the threading
code, which can safely assume that it has the complete threads and no
longer produces any "Message not found" messages.