Make returning replies separately optional
Add a new, undocumented separate_replies=True option to the search API.

If separate_replies=True option is _not_ given to the search API, then
it reverts to its previous behaviour: _do_ include replies in the "rows"
list returned. This is the same behaviour that the search API had befor:
it returns both top-level annotations and replies in the one "rows"
list, but without any guarantee that if some annotations/replies from a
given thread are in the list then all annotations/replies from that
thread will be in it.

If separate_replies=True _is_ given then the API follows the new
behaviour: "rows" contains top-level annotations only, and a separate
"replies" list containing all replies to the annotations in rows is also
inserted into the result.