Tidy up the comments and variable naming relating to Chrome extension install
Having a variable called 'isChrome' and then explaining why
we are trying to work around Edge's spoofing that aims
to deal with exactly this kind of check is silly.

Instead make it more explicit that we are checking for
support for Chrome _extensions_. If Microsoft Edge supports
all the APIs we need in future for that, great, we'll let
them use it.