Move all templates into `templates/` directory
It would be nice if each module could be responsible for their own
templates, but in reality nearly every template forms part of a large
tree of templates, extending base layouts, and so on.

In the interests of consistency, it makes sense to have all templates in
the `h/templates/` directory rather than some being in `h/templates/`
and some being in their respective module directories.