Put all the files for the custom bootstrap build in a single place
Our script and style files are split between h/static/scripts
and h/static/styles in the source tree.

When updating Bootstrap however, these files are part of a single
archive and it is useful to have the config.json file, the CSS
and the JS all in one place.

This also updates the Bootstrap build with styles for
alerts which are still used by the flashbar.