Only call drafts at the correct times
We only need to save an annotation to drafts when the focused group
changes while editing it, and not at other times (e.g. when opening the
editor??).

We only need to restore an annotation from drafts when editing it, and
not at other random times.

Also, don't call updateViewModel() in init() after calling edit(),
because this will overwrite the draft changes that you just restored.