Update AnnotationController on annotationUpdated
Update AnnotationController's view model on the "annotationUpdated" event.

This fixes an issue that "live updates" were not being shown on annotation
cards.

The "annotationUpdated" event is fired by AnnotationController itself after
successfully saving updates to an already-saved annotation. It's also fired by
annotationMapper after received an updated annotation from the server (that was
updated by a different client - the "live updates" feature).

Subscribing to this event allows one function in AnnotationController to handle
updating the view model in both cases.