Run <excerpt> initialization asynchronously
Run initialization of <excerpt> asynchronously so that any
directives inside its children have been resolved by the time
updateContentMaxHeight() is called.

In the context of the annotation list, the maxHeight
style property of the content element was set to 0 for the
annotation quotes because the 'ngRepeat' directive which
creates the DOM elements for the quotes had not run by the time
that updateContentMaxHeight() was called.