Use Date.toLocaleString()'s options arg for more control over date formatting
A downside of using the browser intl APIs to avoid the overhead
of Moment.js is that the output is implementation dependent.

Use toLocaleString()'s options argument where supported
(Firefox, Chrome, Edge, IE 11) for more control over formatting
of dates in annotation last-updated tooltips.

Safari is the only modern browser currently lacking support
for this option. For Safari, use toDateString() and toLocaleDateString()
which generates the most similar format.