Fix excerpt content height when switching from edit back to view mode
The 'max-height' style property was not updated when the <excerpt>'s
enabled state changed from false to true which happened when switching
from annotation edit mode back to view mode.

Use ngStyle to ensure the inline styles applied to the <excerpt>'s
content element are updated as necessary in this case.