Implement a hysteresis threshold in excerpts
Allow the excerpt's contents to exceed the collapsed
height by a threshold set with the 'overflow-hystersis' attribute
before the excerpt's contents are truncated.

This fixes a problem where the contents of an annotation
body could exceed the threshold by only a couple of pixels,
resulting in a 'More' link which resulted in only half a line
of extra text being revealed.