Fix installation of URL polyfill with js-polyfills v0.1.16
js-polyfills v0.1.15 used to install the polyfill on
`this.URL`, where `this` was set to module.exports in
the context of a Browserify bundle. Consequently the result
was available as `require('js-polyfills/url').URL`.

As of v0.1.16 however, it installs the polyfill on `self`,
which is the Window object in the context of the browser
and so `require('js-polyfills/url').URL` is undefined.

Fixes #3064