Support filtering test suites to run using '--grep' argument to gulp
Add support for filtering the tests that are run by Gulp test tasks
using '--grep <pattern>'.

This avoids the need to use the `(describe|it).only` test modifiers to
filter the tests that are run.

The pattern argument is a JS regex expression which is passed directly
to mocha as the 'grep' option.

eg. To start the app test suite in auto-restart mode, but only run tests
for the markdown editor, run:

  gulp test-watch-app --grep markdown