Debounce live-reload notifications
If multiple asset bundles are regenerated in quick succession, including
JS or template files, the client might try to reload whilst some of the
bundles are still being regenerated, resulting in the client failing to
load.

This commit reduces the likelihood of this happening by debouncing
live-reload notifications and makes a note of the issue.