Fix KaTeX font URL paths
The vendor KaTeX CSS file looks for font URLs in the 'fonts/' directory
relative to the location of katex.min.css.

Rewrite these URLs so that they are fetched from '/assets/fonts' instead
of '/assets/styles/fonts'

Note that the CSS references both WOFF and WOFF 2 format fonts but We
only include the .woff version.