Retry requests to /api endpoint
Handle requests to the /api endpoint that fail with network errors in
the same way as failed requests to the /app endpoint by retrying the
request with exponential backoff until it succeeds.

Otherwise, if the request fails then `store.(Annotation|Search)Resource`
are not populated and code that tries to make API requests fails.