Make client test page load embed from same hostname as test page
Previously the live reload server at http://localhost:3000 hardcoded the
embed script to 'http://localhost:5000' which would cause the embed to
fail to load if accessing the live reload server via a different IP.

This happens when accessing the server through a VirtualBox Guest when
using the default networking settings or when accessing it from a say, a
mobile device on the same WiFi network as a laptop serving the dev
environment.

Instead use the same hostname for the live reload WebSocket server
and embed script as the client test page is served from.