Don't update timestamps > one day ago
If an annotation is more than one day old then it gets a timestamp like
"09 May". The next time this timestamp would change would be on
1 Jan 2017 when it would change to "09 May 2016". Rather than setting a
really long timer, just never live update "09 May"-format timestamps.

The case where Hypothesis is left open without reloading a page until
the end of the year, and the case where Hypothesis is opened close to
midnight on Dec 31st, are obscure enough to ignore.

If an annotation is not from the current year then it gets a timestamp
like "14 March 2015". The format of this timestamp will always be the
same, so don't set any timeout to update it.