Save annotation state to drafts whenever annotation component is destroyed
Prior to this unsaved changes were lost when an annotation being edited
was scrolled off-screen and the annotation component was destroyed.

There was already a mechanism in place for persisting unsaved changes
to a temporary store when switching accounts and when switching groups
but this was only triggered when the signed in user changed or the
focused group changed.

This commit unifies the handling of all three cases by listening
for the '$destroy' event which is fired whenever an <annotation>
is about to be destroyed.