Only filter annotations (not replies) against selection
Only annotations, not replies can be selected, so only filter top-level
annotations against the selection map.

This resulted in an issue where 'View N more in conversation' being
shown where 'N' was the number of replies when selecting an annotation
with replies in the page.

See https://github.com/hypothesis/h/pull/3285#issuecomment-218734287