Move canonical sort state to Redux store
This is a step towards getting us to a place where all important state
in the front-end app is managed in the same way - kept in the Redux
store and modified only via actions.

The sort state was duplicated as a property on the $scope and also
stored in the Redux store.

Instead just expose functions on the scope which retrieve this state
directly from the store and dispatch an action to the store when the
user selects a different sort mode.

In addition, the terminology for sorting has been normalised across the
application (sortBy/sortOptions -> sortKey/sortKeysAvailable), and the
ability to sort the stream has been removed, as it was nonsensical and
broken.