Fix display of <ul>, <ol> in annotations
9be804d broke the display of <ul> and <ol> tags in annotation cards by
overriding a bit more than intended -- specifically, it set `list-style:
none;` on <li> in reset.scss when the Compass reset mixins didn't do
that.

This commit fixes the issue by only setting `list-style: none;` on <ul>
and <ol>, which means that the list styles in styled-text.scss can take
effect as originally intended.