Use document fingerprint, not URL, as primary URI for PDFs
Consider the following situation:

- A PDF with fingerprint (F) exists at multiple URLs (U1, U2, ...)
- A user visits the PDF at U1 and annotates it
- A second user visits the PDF at U2

Prior to this commit, due to the way document equivalence for PDFs
functions, the annotations made at U1 will not be loaded, because the
initial search is made for U2 only.

(If the second user subsequently makes annotations at U2, and then
reloads the page, the annotations made at U1 will then show up, as the
association between U2 and F will have been made).

This commit improves the situation here by treating the PDF fingerprint
as the "primary URI" for the document, much as we treat <link
rel="canonical"> URLs as "primary" if we find them in an HTML page.

This means that the initial search will be made for the fingerprint F,
which will match the annotations made at U1.

For more, see:

  https://trello.com/c/DVUemKwi/329-use-pdf-fingerprint-as-primary-search-uri-for-pdfs