Eliminate `vm.form` and `vm.isPrivate`
These variables duplicated the information from the current draft for an
annotation. Instead of restoring a draft when an annotation card is
created and saving it when the annotation card is destroyed, just update
the draft whenever the tags, text or privacy of the annotation is
changed. The text/tags/privacy shown in the card are determined by
`vm.state()` which merges the annotation and the draft.

Removing these variables lets us also remove or simplify the tests that
checked that everything remained in sync.