Fix `deleteIfNewAndEmpty()` being called after annotation was destroyed
The correct fix would be to remove new annotations with an empty draft
in the reducer function that handles new annotations being created. That
is not currently possible because drafts are not stored in the Redux
store. Therefore this commit just manually de-registers the listener
when <annotation> goes out of scope.