Make Annotation and Notes tab bar sticky
Make Annotation and Notes tab bar stick to top of the viewport when the
annotation list is scrolled down.

This relies on support for `position: sticky` which is available in
recent versions of Firefox and Safari*. On browsers where
`position: sticky` is not supported, the tab bar will just scroll
off-screen when scrolling the annotation list down.

Because annotation cards may now be displayed under the tabs, add a
background color, Z index and negative horizontal margin to the tabs
component so that any part of an annotation card under the tab bar is
not visible.

* In Chrome v52 this feature is available but only when the
  'Experimental web platform features' flag is on.