Fix 'time' tests in timezones behind UTC
Several test fixtures have expectations that the "current date", where
"current" is reset to the Unix epoch at the start of the tests, is 1st
Jan 1970. This is true in timezones equal to or ahead of UTC but not the
case for timezones behind UTC.

Resolve the issue by just advancing the fake clock forwards to midnight
on 01/01/1970 if the timezone is behind UTC at the start of the tests.