Remove PhantomJS-specific polyfills (#45)
PhantomJS supports ES5 fully and does not require specific polyfills for
it. This means that we can use exactly the same polyfill script in both
our unit tests and the actual app, which reduces the chances of breakage
due to mismatches.

Additionally, the broken URL constructor could potentially affect other
browsers, so it makes sense to include it in the main polyfill script.