Make the adder a self-contained component
Move adder event handling and UI creation to Adder component, so that
Guest does not need to know about its internal structure or handle
events for it. Instead it just supplies callbacks to invoke when the
Annotate and Highlight buttons are clicked.

This makes it easier to test event handling for the adder and to make
future changes to the UI of the Adder.

This commit removes a piece of logic that stopped propagation of
'mouseup' events from the adder because I couldn't find a clear
explanation in the code or commit history of why that was done.