Replace "firstRun" config option with "openSidebar" and "openLoginForm"
This commit deprecates the confusingly-named "firstRun" embed script
option with two which actually describe the relevant actions triggered.

Unfortunately, at the moment this requires a fairly horrific check that
the string value of a config option is not the literal 'false', due to
the way that the Host code passes configuration from the page frame to
the sidebar frame (using URL query string parameters).