Add a timeout for anchoring in the sidebar
If anchoring an annotation fails to complete within 500ms then assume
that an error occurred during anchoring and that the annotation is
therefore an orphan.

If it later turns out that anchoring just took a long time then the
annotation will simply move from the Orphans tab to the Annotations tab
once anchoring completes.

A timeout within the sidebar app is used rather than relying on the page
code to send back a message if an error occurs during anchoring because
the code that runs in the page context could fail in arbitrary ways due
to interactions with JavaScript on the page.