Feature-flag orphan quote strikethrough on annotation cards
When the orphans feature flag is not set, do not indicate orphans in
annotation cards by striking through their quotes.

The feature flag is hoisted outside of the `isOrphan` call because I
recall that `flagEnabled` was actually quite an expensive call in the
past so we want to avoid invoking it during every digest cycle.