Sort notes tab by date
This commit prevents page notes from being sorted by "Location" (i.e.
document location), a meaningless property for page notes, which would
result in them being sorted unpredictably.

Instead, we define a default sort key and a set of allowable sort keys
for each sidebar tab. When switching between tabs, the sort key is
updated to the default, and the set of allowable sort keys updated
depending on the tab.

This results in page notes being sorted (by default) in ascending order
of their most recent update. Using "updated" rather than "created" is a
bit dubious, but changing this requires a bit of care to avoid strange
behaviour on the stream, so I've left that for another time.

N.B. As implemented, the sort key on each tab will be reset to the
default every time the tab changes. This is less surprising than
preserving the sort key across tab switches, and doesn't involve any
requirement to remember system state across tab switches. We can also
revisit this in the future if it seems wrong.

Fixes #96.