Defer applying updates received via the WebSocket
To avoid having the contents of the sidebar shift in a distracting way
while the user is reading or editing an annotation, do not apply
real-time updates received via the WS immediately but instead alert the
user to the updates via an icon in the top-bar and apply the updates
when the user clicks the button.

This commit implements support for caching updates in the streamer,
showing an icon and applying updates when the user clicks the icon.

Known issues:

 - The icon does not currently match the Refresh icon in the specs

 - The tooltip sits above the icon pointing down, instead of pointing
   up as per the mocks