Fix off-screen empty drafts not being removed when a new annotation is created
The logic to remove new and empty annotations used to live in the
<annotation> component because that was the only part of the code that
had access to the state of unsaved changes.

Since <annotation> instances are only created for on-screen annotations,
this can result in empty drafts not being removed if the empty
annotation is off-screen.

Now that the canonical content of unsaved annotations is stored in the
drafts service, we can move the logic outside of the annotation
component and fix this problem.

Fixes #97